+++ packages/logic-core/src/data/puzzle.ts
@@ -1,5 +1,6 @@
 import { z } from 'zod';
 import GridData from './grid.js';
+import { PuzzleType } from './primitives.js';
 
 export type PuzzleMetadata = {
   /**
@@ -15,10 +16,6 @@ export type PuzzleMetadata = {
    */
   description: string;
   /**
-   * A link to a place to discuss this puzzle. (can be empty)
-   */
-  link: string;
-  /**
    * The difficulty of the puzzle, from 0 to 10. (required)
    *
    * 0 represents an unrated puzzle, 6-10 represent star difficulties.
@@ -31,7 +28,6 @@ export const MetadataSchema = z
     title: z.string().min(1),
     author: z.string().min(1),
     description: z.string(),
-    link: z.string(),
     difficulty: z.number().int().min(0).max(10),
   })
   .strict();
@@ -41,14 +37,13 @@ export const PuzzleSchema = z
     title: z.string().min(1),
     author: z.string().min(1),
     description: z.string(),
-    link: z.string(),
     difficulty: z.number().int().min(0).max(10),
     grid: z.instanceof(GridData),
     solution: z.instanceof(GridData).nullable(),
   })
   .strict();
 
-export type Puzzle = PuzzleMetadata & {
+export type PuzzleData = {
   /**
    * The grid of the puzzle. (required)
    *
@@ -64,3 +59,29 @@ export type Puzzle = PuzzleMetadata & {
    */
   solution: GridData | null;
 };
+
+export type Puzzle = PuzzleMetadata & PuzzleData;
+
+/**
+ * Get the types of a puzzle based on its grid properties. The returned types are ordered by their priority.
+ * The first type is the most important one.
+ */
+export function getPuzzleTypes(puzzle: Puzzle): PuzzleType[] {
+  const types: PuzzleType[] = [];
+  let logic = true;
+  if (puzzle.grid.musicGrid.value) {
+    types.push(PuzzleType.Music);
+    logic = false;
+  }
+  if (puzzle.grid.completePattern.value) {
+    types.push(PuzzleType.Pattern);
+    logic = false;
+  }
+  if (puzzle.grid.underclued.value) {
+    types.push(PuzzleType.Underclued);
+  }
+  if (logic) {
+    types.push(PuzzleType.Logic);
+  }
+  return types;
+}