+++ packages/logic-core/src/index.ts
@@ -19,8 +19,8 @@ import GridData, { NEIGHBOR_OFFSETS } from './data/grid.js';
 import GridConnections from './data/gridConnections.js';
 import GridZones from './data/gridZones.js';
 import Instruction from './data/instruction.js';
-import { COMPARISONS, Color, Comparison, DIRECTIONS, Direction, MajorRule, Mode, ORIENTATIONS, Orientation, State, WRAPPINGS, Wrapping, directionToggle, orientationToggle } from './data/primitives.js';
-import { MetadataSchema, PuzzleSchema } from './data/puzzle.js';
+import { COMPARISONS, Color, Comparison, DIRECTIONS, Direction, MajorRule, Mode, ORIENTATIONS, Orientation, PuzzleType, State, WRAPPINGS, Wrapping, directionToggle, orientationToggle } from './data/primitives.js';
+import { MetadataSchema, PuzzleSchema, getPuzzleTypes } from './data/puzzle.js';
 import BanPatternRule from './data/rules/banPatternRule.js';
 import CellCountPerZoneRule from './data/rules/cellCountPerZoneRule.js';
 import CellCountRule from './data/rules/cellCountRule.js';
@@ -154,6 +154,7 @@ export {
   Mode,
   ORIENTATIONS,
   Orientation,
+  PuzzleType,
   State,
   WRAPPINGS,
   Wrapping,
@@ -161,6 +162,7 @@ export {
   orientationToggle,
   MetadataSchema,
   PuzzleSchema,
+  getPuzzleTypes,
   BanPatternRule,
   CellCountPerZoneRule,
   CellCountRule,